% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{app_server}
\alias{app_server}
\title{Main Shiny server function for secuTrialRshiny}
\usage{
app_server(input, output)
}
\arguments{
\item{input}{session's input object}

\item{output}{session's output object}
}
\description{
Calls all modules required by app_ui, and passes on relevant data.
}
\seealso{
\code{\link{app_ui}}, \code{\link{run_shiny}}
}
