% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{app_ui}
\alias{app_ui}
\title{Main shiny UI function for secuTrialRshiny}
\usage{
app_ui()
}
\value{
shiny.tag list object containing the dashboard page
}
\description{
Controls the general appearance, header, sidebar, body, tabs with modules, etc.
}
\seealso{
\code{\link{app_srv}}, \code{\link{run_shiny}}
}
