% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/com_footer.R
\name{com_footer_UI}
\alias{com_footer_UI}
\title{A shiny component generating an HTML footer line}
\usage{
com_footer_UI(footer_info)
}
\arguments{
\item{footer_info}{character containing a string to be encapsulated by the "footer" tag}
}
\value{
html string with a "footer" tag
}
\description{
This is used to generate a footer line containing info about the
file loaded into the shiny app via read_secuTrialR()
}
\seealso{
\code{\link{app_srv}}, \code{\link{run_shiny}}
}
