% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloader.R
\name{downloader}
\alias{downloader}
\title{secuTrialdata converter and download handler}
\usage{
downloader(file_name, format, sT_export)
}
\arguments{
\item{file_name}{string containing a file name, including the extension}

\item{format}{string containing a format identifier for secuTrialR::write_secuTrial()}

\item{sT_export}{secuTrialdata object generated e.g. with secuTrialR::read_secuTrial()}
}
\description{
This is an internal helper function allowing conversion and download of secuTrialdata object
using a shiny::downloadHandler(). Supported file formats are the ones listed in secuTrialR::write_secuTrial().
}
\seealso{
\code{\link{mod_export_srv}}
}
