% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_modules.R
\name{get_modules}
\alias{get_modules}
\title{Retrieves a list of all module aliases}
\usage{
get_modules()
}
\value{
list of strings containing module aliases
}
\description{
Module aliases retrieved via get_modules() are used to define module IDs in shiny::callModule(),
and are also used as tabnames for the sidebar, and for the shiny UI label and id.
}
\seealso{
\code{\link{app_ui}}, \code{\link{app_srv}}
}
