% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_style.R
\name{get_style}
\alias{get_style}
\title{Retrieves a list of style settings}
\usage{
get_style(style = c("scto", "dkf"))
}
\arguments{
\item{style}{string defining the css style to be used for the shiny app}
}
\value{
list of strings containing style settings
}
\description{
Retrieves a list of style settings
}
\seealso{
\code{\link{app_ui}}, \code{\link{app_srv}}
}
