% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_codebook.R
\name{mod_codebook}
\alias{mod_codebook}
\title{Shiny module server function for codebook display}
\usage{
mod_codebook(input, output, session, sT_export)
}
\arguments{
\item{input}{session's input object}

\item{output}{session's output object}

\item{session}{session object environment}

\item{sT_export}{secuTrialdata object generated e.g. with secuTrialR::read_secuTrial()}
}
\description{
This function represents a shiny dashboard server module that allows users to
display a codebook of a secuTrialRdata object.
}
\seealso{
\code{\link{mod_codebook_UI}}
}
