% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_codebook.R
\name{mod_codebook_UI}
\alias{mod_codebook_UI}
\title{Shiny module UI function for codebook display}
\usage{
mod_codebook_UI(id, label)
}
\arguments{
\item{id}{string containing a namespace identifier}

\item{label}{string to be used as sidebar tab label}
}
\value{
shiny.tag list object containing the tab item content
}
\description{
This function represents a shiny dashboard UI module that allows users to
display a codebook of a secuTrialRdata object.
}
\seealso{
\code{\link{mod_codebook_srv}}
}
