% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_export.R
\name{mod_export}
\alias{mod_export}
\title{Shiny module server function for data conversion and export}
\usage{
mod_export(input, output, session, sT_export)
}
\arguments{
\item{input}{session's input object}

\item{output}{session's output object}

\item{session}{session object environment}

\item{sT_export}{secuTrialdata object generated e.g. with secuTrialR::read_secuTrial()}
}
\description{
This function represents a shiny dashboard server module that allows users to
convert secuTrialRdata to STATA, SAS or SPSS compatible formats, and download
them as a zip archive.
}
\seealso{
\code{\link{mod_export_UI}}
}
