% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_export.R
\name{mod_export_UI}
\alias{mod_export_UI}
\title{Shiny module UI function for data conversion and export}
\usage{
mod_export_UI(id, label)
}
\arguments{
\item{id}{string containing a namespace identifier}

\item{label}{string to be used as sidebar tab label}
}
\value{
shiny.tag list object containing the tab item content
}
\description{
This function represents a shiny dashboard UI module that allows users to
convert secuTrialRdata to STATA, SAS or SPSS compatible formats, and download
them as a zip archive.
}
\seealso{
\code{\link{mod_export}}
}
