% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_formcomplete.R
\name{mod_formcomplete}
\alias{mod_formcomplete}
\title{Shiny module server function for form completeness monitoring}
\usage{
mod_formcomplete(input, output, session, sT_export)
}
\arguments{
\item{input}{session's input object}

\item{output}{session's output object}

\item{session}{session object environment}

\item{sT_export}{secuTrialdata object generated e.g. with secuTrialR::read_secuTrial()}
}
\description{
This function represents a shiny dashboard server module that allows users to
view a form completeness table.
}
\seealso{
\code{\link{mod_formcomplete_UI}}
}
