% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_formcomplete.R
\name{mod_formcomplete_UI}
\alias{mod_formcomplete_UI}
\title{Shiny module UI function for form completeness monitoring}
\usage{
mod_formcomplete_UI(id, label)
}
\arguments{
\item{id}{string containing a namespace identifier}

\item{label}{string to be used as sidebar tab label}
}
\value{
shiny.tag list object containing the tab item content
}
\description{
This function represents a shiny dashboard UI module that allows users to
view a form completeness table.
}
\seealso{
\code{\link{mod_formcomplete}}
}
