% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_monitorcn.R
\name{mod_monitorcn}
\alias{mod_monitorcn}
\title{Shiny module server function for casenode monitoring}
\usage{
mod_monitorcn(input, output, session, sT_export)
}
\arguments{
\item{input}{session's input object}

\item{output}{session's output object}

\item{session}{session object environment}

\item{sT_export}{secuTrialdata object generated e.g. with secuTrialR::read_secuTrial()}
}
\description{
This function represents a shiny dashboard server module that allows users to
generate a list of random participants for monitoring.
The list can be downloaded as a csv file.
}
\seealso{
\code{\link{mod_monitorcn_UI}}
}
