% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_monitorcn.R
\name{mod_monitorcn_UI}
\alias{mod_monitorcn_UI}
\title{Shiny module UI function for casenode monitoring}
\usage{
mod_monitorcn_UI(id, label)
}
\arguments{
\item{id}{string containing a namespace identifier}

\item{label}{string to be used as sidebar tab label}
}
\value{
shiny.tag list object containing the tab item content
}
\description{
This function represents a shiny dashboard UI module that allows users to
generate a list of random participants for monitoring.
The list can be downloaded as a csv file.
}
\seealso{
\code{\link{mod_monitorcn_srv}}
}
