% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_recruitplot.R
\name{mod_recruitplot}
\alias{mod_recruitplot}
\title{Shiny module server function for recruitment plot display}
\usage{
mod_recruitplot(input, output, session, sT_export)
}
\arguments{
\item{input}{session's input object}

\item{output}{session's output object}

\item{session}{session object environment}

\item{sT_export}{secuTrialdata object generated e.g. with secuTrialR::read_secuTrial()}
}
\description{
This function represents a shiny dashboard server module that allows users to
view a secuTrialR recruitment plot.
}
\seealso{
\code{\link{mod_recruitplot_UI}}
}
