% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_recruitplot.R
\name{mod_recruitplot_UI}
\alias{mod_recruitplot_UI}
\title{Shiny module UI function for recruitment plot display}
\usage{
mod_recruitplot_UI(id, label)
}
\arguments{
\item{id}{string containing a namespace identifier}

\item{label}{string to be used as sidebar tab label}
}
\value{
shiny.tag list object containing the tab item content
}
\description{
This function represents a shiny dashboard UI module that allows users to
view a secuTrialR recruitment plot.
}
\seealso{
\code{\link{mod_recruitplot}}
}
