% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_recruittable.R
\name{mod_recruittable_srv}
\alias{mod_recruittable_srv}
\title{Shiny module server function for recruitment table display}
\usage{
mod_recruittable_srv(input, output, session, sT_export, vals_upload)
}
\arguments{
\item{input}{session's input object}

\item{output}{session's output object}

\item{session}{session object environment}

\item{sT_export}{secuTrialdata object generated e.g. with secuTrialR::read_secuTrial()}

\item{vals_upload}{reactivevalues list containing the output of \code{\link{mod_upload_srv}}}
}
\description{
This function represents a shiny dashboard server module that allows users to
view a secuTrialR recruitment table.
}
\seealso{
\code{\link{mod_recruittable_UI}}
}
