% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_recruittable.R
\name{mod_recruittable_UI}
\alias{mod_recruittable_UI}
\title{Shiny module UI function for recruitment table display}
\usage{
mod_recruittable_UI(id, label)
}
\arguments{
\item{id}{string containing a namespace identifier}

\item{label}{string to be used as sidebar tab label}
}
\value{
shiny.tag list object containing the tab item content
}
\description{
This function represents a shiny dashboard UI module that allows users to
view a secuTrialR recruitment table.
}
\seealso{
\code{\link{mod_recruittable_srv}}
}
