% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_upload.R
\name{mod_upload_srv}
\alias{mod_upload_srv}
\title{Shiny module server function for data upload}
\usage{
mod_upload_srv(input, output, session, sT_export)
}
\arguments{
\item{input}{session's input object}

\item{output}{session's output object}

\item{session}{session object environment}

\item{sT_export}{secuTrialdata object generated e.g. with secuTrialR::read_secuTrial()}
}
\description{
This function represents a shiny dashboard server module that allows users to
upload secuTrial export .zip to the app. This is a prerequisite for using other
modules of secuTrialRshiny.
}
\seealso{
\code{\link{mod_upload_UI}}
}
