% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_upload.R
\name{mod_upload_UI}
\alias{mod_upload_UI}
\title{Shiny module UI function for data upload}
\usage{
mod_upload_UI(id, label)
}
\arguments{
\item{id}{string containing a namespace identifier}

\item{label}{string to be used as sidebar tab label}
}
\value{
shiny.tag list object containing the tab item content
}
\description{
This function represents a shiny dashboard UI module that allows users to
upload secuTrial export .zip to the app. This is a prerequisite for using other
modules of secuTrialRshiny.
}
\seealso{
\code{\link{mod_upload}}
}
