% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_visitplan.R
\name{mod_visitplan_UI}
\alias{mod_visitplan_UI}
\title{Shiny module UI function for visit plan display}
\usage{
mod_visitplan_UI(id, label)
}
\arguments{
\item{id}{string containing a namespace identifier}

\item{label}{string to be used as sidebar tab label}
}
\value{
shiny.tag list object containing the tab item content
}
\description{
This function represents a shiny dashboard UI module that allows users to
view a plot of the visit plan.
}
\seealso{
\code{\link{mod_visitplan_srv}}
}
